+++ src/main/java/org/opentripplanner/api/model/ApiVehicleRentalStation.java
@@ -20,8 +20,8 @@ public class ApiVehicleRentalStation {
     public List<String> networks = null;
 
     /**
-     * Whether this station is static (usually coming from OSM data) or a real-time source. If no real-time data, users should take
-     * bikesAvailable/spacesAvailable with a pinch of salt, as they are always the total capacity divided by two. Only the total is meaningful.
+     * Whether this station has real-time data available currently. If no real-time data, users should take
+     * bikesAvailable/spacesAvailable with a pinch of salt, as they are always the total capacity divided by two.
      */
     public boolean realTimeData = true;