+++ src/main/java/org/opentripplanner/model/plan/Place.java
@@ -5,6 +5,7 @@ import org.opentripplanner.model.StopLocation;
 import org.opentripplanner.model.WgsCoordinate;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.base.ToStringBuilder;
+import org.opentripplanner.routing.vehicle_rental.VehicleRentalPlace;
 
 /** 
 * A Place is where a journey starts or ends, or a transit stop along the way.
@@ -45,9 +46,9 @@ public class Place {
     public VertexType vertexType;
 
     /**
-     * In case the vertex is of type Bike sharing station.
+     * In case the vertex is of type vehicle sharing station.
      */
-    public FeedScopedId bikeShareId;
+    public VehicleRentalPlace vehicleRentalStation;
 
     public Place(Double lat, Double lon, String name) {
         this.name = name;
@@ -99,7 +100,7 @@ public class Place {
                 .addNum("stopIndex", stopIndex)
                 .addNum("stopSequence", stopSequence)
                 .addEnum("vertexType", vertexType)
-                .addObj("bikeShareId", bikeShareId)
+                .addObj("vehicleRentalId", vehicleRentalStation)
                 .toString();
     }
 }