+++ src/main/java/org/opentripplanner/updater/vehicle_rental/VehicleRentalDataSource.java
@@ -2,8 +2,7 @@ package org.opentripplanner.updater.vehicle_rental;
 
 import java.util.List;
 
-import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.vehicle_rental.VehicleRentalStation;
+import org.opentripplanner.routing.vehicle_rental.VehicleRentalPlace;
 
 /**
  * TODO clarify thread safety.
@@ -20,7 +19,7 @@ public interface VehicleRentalDataSource {
     /**
      * @return a List of all currently known vehicle rental stations. The updater will use this to update the Graph.
      */
-    List<VehicleRentalStation> getStations();
+    List<VehicleRentalPlace> getStations();
 
     /**
      * @see org.opentripplanner.updater.GraphUpdater#setup